+++ docs/CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to Data Machine Code will be documented in this file.
 
+## [0.59.3] - 2026-08-22
+
+### Fixed
+- atomically reconcile runtime symlinks
+
+## [0.59.2] - 2026-08-22
+
+### Fixed
+- harden demand calibration
+- calibrate bootstrap demand
+
+## [0.59.1] - 2026-08-22
+
+### Fixed
+- lock cleanup revalidation
+- revalidate cleanup under repo lock
+- protect capacity remediation previews
+- preserve capacity remediation evidence
+- harden capacity remediation
+- remediate refused capacity admission
+
 ## [0.59.0] - 2026-08-22
 
 ### Added